Moved several enums from Login to Mabi
