+++ src/LoginServer/Network/Handlers/Deletion.cs
@@ -5,6 +5,7 @@ using System;
 using Aura.Login.Database;
 using Aura.Shared.Network;
 using Aura.Mabi.Network;
+using Aura.Mabi.Const;
 
 namespace Aura.Login.Network.Handlers
 {
